// GestionStock/src/app/services/openapi/model/adresse-request.ts
/**
 * Gestion de stock REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AdresseRequest { 
    adresse1: string;
    adresse2?: string;
    ville: string;
    codePostale?: string;
    pays: string;
}