// GestionStock/src/app/services/openapi/model/client-request.ts
/**
 * Gestion de stock REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdresseRequest } from './adresse-request';


export interface ClientRequest { 
    id?: number;
    nom: string;
    numeroCNI: string;
    sexe: ClientRequest.SexeEnum;
    numeroTel: string;
    email?: string;
    adresse: AdresseRequest;
}
export namespace ClientRequest {
    export type SexeEnum = 'MASCULIN' | 'FEMININ';
    export const SexeEnum = {
        Masculin: 'MASCULIN' as SexeEnum,
        Feminin: 'FEMININ' as SexeEnum
    };
}