// GestionStock/src/app/services/openapi/model/fournisseur-request.ts
/**
 * Gestion de stock REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdresseRequest } from './adresse-request';


export interface FournisseurRequest { 
    id?: number;
    nom: string;
    numeroCNI?: string;
    sexe?: FournisseurRequest.SexeEnum;
    numeroTel: string;
    email?: string;
    adresse: AdresseRequest;
}
export namespace FournisseurRequest {
    export type SexeEnum = 'MASCULIN' | 'FEMININ';
    export const SexeEnum = {
        Masculin: 'MASCULIN' as SexeEnum,
        Feminin: 'FEMININ' as SexeEnum
    };
}